requirements.txt
python-dotenv
streamlit
langchain
openai
snowflake-connector-python
pandas
langchain-community
